Fix unable to pair when using email method
Every time the welcome view was loaded, it was re-generating a new pairing token. This meant that if you copied the pairing token and opened another app to send the token (e.g. Email app), once you returned to the Bisq app it would have generated a new token, invalidating the old one and unable to decrypt the messages.