Fix crash when opening offer alert notifications
When opening an offer alert notification, it was loading the BTC average price from bitcoinaverage.com, but would crash if the response was not as expected - attempting to interact with a null object. Bitcoinaverage have since changed to a subscription based model and is no longer free, so was returning an error response.

Since loading the BTC average price over clear net could be considered a privacy concern, we no longer attempt to retrieve it.