Automatically load success view for simulator in SendToken view
When using the simulator, automatically jump to the success screen shortly after loading the send token view as opposed to the QR view. This allows for testing the QR code in the simulator without it jumping to the success view. Also, increased the delay to allow for more time on the send token view.